% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnsignature_functions.R
\name{calculateSumOfPosteriors}
\alias{calculateSumOfPosteriors}
\title{Calculate sum of posteriors}
\usage{
calculateSumOfPosteriors(
  CN_feature,
  components,
  name,
  rowIter = 1000,
  cores = 1
)
}
\arguments{
\item{CN_feature}{feature of the copy-number profiles}

\item{components}{components of copy-number signatures}

\item{name}{name of the component}

\item{rowIter}{iterations for calculation of sum of posteriors. Default is 1000}

\item{cores}{number of cores to use. Default is 1}
}
\value{
Sum of posteriors per component
}
\description{
Helper function for generateSampleByComponentMatrix.
Approach developed by Geoff Macintyre et al. 2018
}
