% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CCLE_metadata}
\alias{CCLE_metadata}
\title{Metadata of the CCLE dataset}
\format{
his data frame has 1,046 rows (cell lines) and the following
13 columns:
\describe{
\item{CCLE.name}{CCLE id}
\item{Cell.line.primary.name}{cell line id}
\item{Cell.line.aliases}{}
\item{Gender}{"M" or "F"}
\item{Site.Primary}{primary site of tumour}
\item{Histology}{histology type}
\item{Hist.Subtype1}{histology subtype}
\item{Notes}{}
\item{Source}{}
\item{Expression.arrays}{file id}
\item{SNP.arrays}{file id}
\item{Oncomap}{}
\item{Hybrid.Capture.Sequencing}{}
}
}
\usage{
CCLE_metadata
}
\description{
This data includes metadata of cell lines included in the
Cancer Cell Line Encyclopaedia (CCLE) project.
}
\keyword{datasets}
