% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cells_mapping}
\alias{cells_mapping}
\title{Information of dataset from where genomic data of cell lines are downloaded}
\format{
This data frame has 1,946 rows and 8 columns.
}
\usage{
cells_mapping
}
\description{
This data includes the study where cell lines are included.
}
\keyword{datasets}
