% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cells_segcn}
\alias{cells_segcn}
\title{Precomputed dataset of 1,417 human cancer cell line profiles.}
\format{
This data frame has 101,637 rows and 5 columns.
\describe{
\item{chromosome}{}
\item{start}{}
\item{end}{}
\item{segVal}{raw copy numbers}
\item{sample}{name of the sample}
}
}
\usage{
cells_segcn
}
\description{
This dataset contains the absolute copy number profiles of
1,417 human cancer cell lines. Profiles are derived using ASCAT.
}
\keyword{datasets}
