% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{CNconvert}
\alias{CNconvert}
\title{A helper function for CNPlot_events}
\usage{
CNconvert(e, base)
}
\arguments{
\item{e}{segment table with absolute copy numbers}

\item{base}{sizes of chromosomes}
}
\value{
segment table prepared for plotting
}
\description{
This function prepare profiles for plotting
}
\examples{
exp_cell=cells_segcn[cells_segcn$sample=="22RV1",]
base=CNpare:::chr_sizes
base$length <- 1e-6 * base$length #Convert sizes to Mb
base$offset <- cumsum(base$length) - base$length #Calculate the interval range of each chr
events<-CNconvert(exp_cell,base)
}
