% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNpare-package.R
\docType{package}
\name{CNpare-package}
\alias{CNpare}
\alias{CNpare-package}
\title{CNpare: Comparing tumour copy number profiles}
\description{
CNpare can be used to: 
    1) calculate similarities between copy-number profiles
    2) quantify the extent of genome differences between copy-number profiles
    3) compare profiles based on copy-number signatures
}
\author{
\strong{Maintainer}: Barbara Hernando \email{bhernando@cnio.es}

}
\keyword{internal}
