% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnsignature_functions.R
\name{generateSampleByComponentMatrix}
\alias{generateSampleByComponentMatrix}
\title{Get components per sample}
\usage{
generateSampleByComponentMatrix(
  CN_features,
  all_components = NULL,
  cores = 1,
  rowIter = 1000,
  subcores = 2
)
}
\arguments{
\item{CN_features}{list with features per sample}

\item{all_components}{components of copy-number signatures to obtain}

\item{cores}{number of cores to use. Default is 1}

\item{rowIter}{iterations for calculation of sum of posteriors}

\item{subcores}{number of subcores to use. Default is 2}
}
\value{
A matrix with components per sample
}
\description{
This function get components per sample.
Approach developed by Geoff Macintyre et al. 2018
}
