% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessing_functions.R
\name{getBinsStartsEnds}
\alias{getBinsStartsEnds}
\title{Get genomic positions of bins}
\usage{
getBinsStartsEnds(window = 5e+05, chr, lengthChr)
}
\arguments{
\item{window}{numeric variable with the bin size. Default is 500kb}

\item{chr}{character variable with the chromosome name}

\item{lengthChr}{numeric variable with the chromosome length}
}
\value{
list with genomic positions of each bin
}
\description{
This function split genome into evenly sized bins
}
\examples{
posBins <- lapply(seq_len(22),function(chr)
    getBinsStartsEnds(window=500000, chr, lengthChr[chr]))
}
