% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnsignature_functions.R
\name{getBPnum}
\alias{getBPnum}
\alias{getOscilation}
\title{Get counts of brekpoints per 10Mb}
\usage{
getBPnum(abs_profiles, chrlen)

getOscilation(abs_profiles, chrlen)
}
\arguments{
\item{abs_profiles}{absolute copy-number profiles}

\item{chrlen}{chromosome length}
}
\value{
Matrix with counts of break points per 10Mb per sample

Matrix with length of chains of oscilating copy numbers
}
\description{
Helper function for extractCopynumberFeatures.
Approach developed by Geoff Macintyre et al. 2018

Helper function for extractCopynumberFeatures.
Approach developed by Geoff Macintyre et al. 2018
}
\examples{
profiles <- getProfiles(segcn=cells_segcn,
    samples=unique(cells_segcn$sample)[1:4])
osCN<-getOscilation(profiles,CNpare:::chrlen)
}
