% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnsignature_functions.R
\name{getChangepointCN}
\alias{getChangepointCN}
\title{Get difference in copy number between adjacent segments}
\usage{
getChangepointCN(abs_profiles)
}
\arguments{
\item{abs_profiles}{absolute copy-number profiles}
}
\value{
Matrix with difference in copy number between adjacent segments
}
\description{
Helper function for extractCopynumberFeatures.
Approach developed by Geoff Macintyre et al. 2018
}
