% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessing_functions.R
\name{getCINProfiles}
\alias{getCINProfiles}
\title{Get profiles with detectable CIN}
\usage{
getCINProfiles(segcn, samples)
}
\arguments{
\item{segcn}{segment table of copy numbers of all samples}

\item{samples}{vector with sample names}
}
\value{
segment table of copy numbers of samples with detectable CIN
}
\description{
This function select samples with detectable CIN
}
\examples{
cin.profiles <- getCINProfiles(segcn=cells_segcn,
    samples=unique(cells_segcn$sample)[1:10])
}
