% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnsignature_functions.R
\name{getCN}
\alias{getCN}
\title{Get copy numbers}
\usage{
getCN(abs_profiles)
}
\arguments{
\item{abs_profiles}{absolute copy-number profiles}
}
\value{
Matrix with copy numbers
}
\description{
Helper function for extractCopynumberFeatures.
Approach developed by Geoff Macintyre et al. 2018
}
\examples{
profiles<-getProfiles(segcn=cells_segcn,
    samples=unique(cells_segcn$sample)[1:4])
copynumber<-getCN(profiles)
}
