% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessing_functions.R
\name{getCNbins.bin}
\alias{getCNbins.bin}
\title{Get copy-number in a bin}
\usage{
getCNbins.bin(b, pb, overlap, samples)
}
\arguments{
\item{b}{bin number}

\item{pb}{matrix with genomic positions of bins}

\item{overlap}{table with overlaps between bin and segment table}

\item{samples}{vector with sample names}
}
\value{
a list of copy-number values per bin
}
\description{
This is a helper function for transforming segment tables to bin tables
}
\examples{
posBins <- lapply(1,function(chr)
    getBinsStartsEnds(window=500000, chr, lengthChr[chr]))
pb=data.table::rbindlist(posBins)
samp=unique(cells_segcn$sample)[1]
data=cells_segcn[cells_segcn$sample==samp,]
data<-data.table::as.data.table(data[,c(5,4,1:3)])
data.table::setkey(pb, chromosome, start, end)
overlap=as.data.frame(data.table::foverlaps(data,pb))
cn_bin <- getCNbins.bin(b=2,pb,overlap,samples=samp)
}
