% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessing_functions.R
\name{getCNbins}
\alias{getCNbins}
\title{Get copy-number per bins}
\usage{
getCNbins(posBins, data, samples)
}
\arguments{
\item{posBins}{list with genomic positions of bins. Each list contains
data from each chromosome. Obtained from getBinsStartsEnds}

\item{data}{segment table of copy numbers of all samples}

\item{samples}{vector with sample names}
}
\value{
bin table of copy numbers of all samples
}
\description{
This function transform segment tables to bin tables
}
