% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifydiff_functions.R
\name{getDifference}
\alias{getDifference}
\title{Calculation of the extent of differences between two profiles}
\usage{
getDifference(events, events_2, method = "non-normalized")
}
\arguments{
\item{events}{segment table with absolute copy numbers of one sample}

\item{events_2}{segment table with absolute copy numbers of the other sample}

\item{method}{method used for calculating difference. Options are "normalized"
or "non-normalized" by ploidy status. Default is "non-normalized"}
}
\value{
\% genome difference between two copy-number profiles
}
\description{
This function computes the \% genome difference
}
\examples{
diff<-getDifference(events=cells_segcn[cells_segcn$sample=="OVKATE",],
    events_2=cells_segcn[cells_segcn$sample=="OV-90",])
}
