% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtool_functions.R
\name{getInputmatrix}
\alias{getInputmatrix}
\title{Get input data as matrix}
\usage{
getInputmatrix(dat)
}
\arguments{
\item{dat}{bin-level copy numbers}
}
\value{
matrix with copy numbers of all samples
}
\description{
This function prepare data for comparisons
}
