% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifydiff_functions.R
\name{getNewSegments}
\alias{getNewSegments}
\title{Create new segments for unifying profiles in a specific chromosome}
\usage{
getNewSegments(ref, exp, window)
}
\arguments{
\item{ref}{dataframe with segments of the reference sample}

\item{exp}{dataframe with segments of the experimental model}

\item{window}{boundary margin for smoothing segments. Default is 100000}
}
\value{
dataframe with new segments in a specific chromosome
}
\description{
This function creates unified and collapse segments within a window distance.
This function in used in unifySegments function, where a table with new segments per
chromosome is generated for generating a unified copy number table.
}
\examples{
posSeg=cells_segcn[cells_segcn$sample=="OVKATE",]
data=cells_segcn[cells_segcn$sample=="OV-90",]
new_seg<-getNewSegments(ref=posSeg[posSeg$chromosome==1,],
    exp=data[data$chromosome==1,],window=100000)
}
