% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifydiff_functions.R
\name{getPloidy}
\alias{getPloidy}
\title{Calculation of ploidy status}
\usage{
getPloidy(events)
}
\arguments{
\item{events}{dataframe with segments of the sample}
}
\value{
ploidy status of the sample
}
\description{
This function calculate the overall ploidy of samples by computing
the weighted mean copy number values of all segments
}
\examples{
exp_cell=cells_segcn[cells_segcn$sample=="22RV1",]
ploidy<-getPloidy(exp_cell)
}
