% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessing_functions.R
\name{getProfiles}
\alias{getProfiles}
\title{Add profiles in a list}
\usage{
getProfiles(segcn, samples)
}
\arguments{
\item{segcn}{segment table of copy numbers of all samples}

\item{samples}{vector with sample names}
}
\value{
list of profiles
}
\description{
This function includes all profiles in a list
}
\examples{
profiles <- getProfiles(segcn=cells_segcn,
    samples=unique(cells_segcn$sample))
}
