% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnsignature_functions.R
\name{getSampNames}
\alias{getSampNames}
\title{Get sample names}
\usage{
getSampNames(abs_profiles)
}
\arguments{
\item{abs_profiles}{absolute copy-number profiles}
}
\value{
name of each sample
}
\description{
Helper function for getting the name of samples.
Approach developed by Geoff Macintyre et al. 2018
}
