% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessing_functions.R
\name{getSegRounded}
\alias{getSegRounded}
\title{Round copy-number values of segments}
\usage{
getSegRounded(profiles)
}
\arguments{
\item{profiles}{list of profiles}
}
\value{
list of profiles with copy numbers rounded
}
\description{
This function rounds copy numbers to the closest integer
}
\examples{
profiles <- getProfiles(segcn=cells_segcn,
    samples=unique(cells_segcn$sample)[1:10])
rounded.profiles<-getSegRounded(profiles)
}
