% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnsignature_functions.R
\name{getSegsize}
\alias{getSegsize}
\title{Get length of segments}
\usage{
getSegsize(abs_profiles)
}
\arguments{
\item{abs_profiles}{absolute copy-number profiles}
}
\value{
Matrix with segment sizes per sample
}
\description{
Helper function for extractCopynumberFeatures.
Approach developed by Geoff Macintyre et al. 2018
}
