% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnsignature_functions.R
\name{getSegTable}
\alias{getSegTable}
\title{Get segmented table of each sample}
\usage{
getSegTable(x)
}
\arguments{
\item{x}{each object included}
}
\value{
segment table per sample
}
\description{
Helper function for getting segment table with copy numbers.
Approach developed by Geoff Macintyre et al. 2018
}
\examples{
profiles<-getProfiles(segcn=cells_segcn,
    samples=unique(cells_segcn$sample)[1:4])
segTab<-profiles[[1]]
}
