% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtool_functions.R
\name{getSimilarities}
\alias{getSimilarities}
\title{Get similarity metrics}
\usage{
getSimilarities(dat1, dat2, method = "all")
}
\arguments{
\item{dat1}{matrix with bin-level copy numbers per sample to test}

\item{dat2}{matrix with bin-level copy numbers per model to compare}

\item{method}{method to use for testing similarity. Default is all}
}
\value{
dataframe with similarity metrics for all comparisons
}
\description{
This function get similarity metrics for comparisons
}
