% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtool_functions.R
\name{getSimilarities.sample}
\alias{getSimilarities.sample}
\title{Get similarity metrics per sample}
\usage{
getSimilarities.sample(s, dat1, dat2, method, pvalue)
}
\arguments{
\item{s}{sample name}

\item{dat1}{matrix with bin-level copy numbers per sample to test}

\item{dat2}{matrix with bin-level copy numbers per model to compare}

\item{method}{method to use for testing similarity. Default is \emph{all} methods.}

\item{pvalue}{a logical value to indicate if the empirical pvalue is calculated or not}
}
\value{
dataframe with similarity metrics for all comparisons of the sample
}
\description{
This function get similarity metrics for comparisons per sample
}
\examples{
posBins<-lapply(seq_len(22),function(chr)
    getBinsStartsEnds(window=500000, chr, lengthChr[chr]))
ccle_cn<-getCNbins(posBins=posBins,data=cells_segcn,samples=unique(cells_segcn$sample)[1:2])
exp_cell<-as.matrix(ccle_cn[,1])
colnames(exp_cell)<-unique(cells_segcn$sample)[1]
measures<-getSimilarities.sample(s=colnames(exp_cell),dat1=exp_cell,
    dat2=ccle_cn,method="pearson",pvalue=FALSE)
}
