% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtool_functions.R
\name{getTopHit}
\alias{getTopHit}
\title{Get top hits}
\usage{
getTopHit(samples, measure, method = "all")
}
\arguments{
\item{samples}{vector with name of samples}

\item{measure}{dataframe with similarity metrics}

\item{method}{similarity metric to use for ordering . Default is all}
}
\value{
dataframe with similarity metrics for all comparisons
}
\description{
This function get the top hit for each comparison per sample
}
