% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifydiff_functions.R
\name{getUnifiedCN}
\alias{getUnifiedCN}
\title{Get copy number values of each new unified segment from a chromosome}
\usage{
getUnifiedCN(new_seg, data)
}
\arguments{
\item{new_seg}{dataframe defining the new unified segments in a specific chromosome}

\item{data}{dataframe with original segments of the sample in a specific chromosome}
}
\value{
dataframe with unified segment tables
}
\description{
This function generates a vector with copy number values
of each new segment, which has been defined using getNewSegments function.
This function in used in unifySegments function, where unified copy number values
in each new chromosome segments is obtained to generate a unified copy number table
}
\examples{
posSeg=cells_segcn[cells_segcn$sample=="OVKATE",]
data=cells_segcn[cells_segcn$sample=="OV-90",]
new_seg<-getNewSegments(ref=posSeg[posSeg$chromosome==1,],
    exp=data[data$chromosome==1,],window=250000)
new_cn<-getUnifiedCN(new_seg,posSeg[posSeg$chromosome==1,])
}
