% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lengthChr}
\alias{lengthChr}
\title{Length of chromosomes}
\format{
vector with the size of chromosomes.
}
\usage{
lengthChr
}
\description{
This dataset contains the length of chromosomes (hg19 version)
}
\keyword{datasets}
