% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnsignature_functions.R
\name{lower_norm}
\alias{lower_norm}
\title{Get lower normalized value}
\usage{
lower_norm(x, sig_thresh = 0.01)
}
\arguments{
\item{x}{value}

\item{sig_thresh}{threshold for each signature. Default is 0.01}
}
\value{
lower normalized value
}
\description{
Helper function for normaliseMatrix.
Approach developed by Geoff Macintyre et al. 2018
}
