% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnsignature_functions.R
\name{normaliseMatrix}
\alias{normaliseMatrix}
\title{Normalize matrix}
\usage{
normaliseMatrix(signature_by_sample, sig_thresh = 0.01)
}
\arguments{
\item{signature_by_sample}{matrix with signature levels per sample}

\item{sig_thresh}{threshold for each signature. Default is 0.01}
}
\value{
matrix with normalized activity of each signature per sample
}
\description{
Helper function for quantifySignatures.
Approach developed by Geoff Macintyre et al. 2018
}
