% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtool_functions.R
\name{pair_cosine}
\alias{pair_cosine}
\title{Calculation of cosine similarity}
\usage{
pair_cosine(cell, ccle, ccle.name)
}
\arguments{
\item{cell}{vector with bin-level copy number values of one sample}

\item{ccle}{vector with bin-level copy number values of model}

\item{ccle.name}{name of model}
}
\value{
Cosine similarity
}
\description{
This function calculates the Cosine similarity
}
