% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtool_functions.R
\name{pair_pearson}
\alias{pair_pearson}
\title{Perform a pearson correlation test}
\usage{
pair_pearson(cell, ccle, ccle.name)
}
\arguments{
\item{cell}{vector with bin-level copy number values of one sample}

\item{ccle}{vector with bin-level copy number values of model}

\item{ccle.name}{name of model}
}
\value{
Pearson's r
}
\description{
This function calculates the Pearson's r
}
