% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_diffdensity}
\alias{plot_diffdensity}
\title{Density plot showing the distribution of genome differences}
\usage{
plot_diffdensity(diff)
}
\arguments{
\item{diff}{dataframe with cell names 'cellid' and genome differences 'percDiff'}
}
\value{
A density plot with the distribution. The min and max value is also reported
}
\description{
This function draws the extent of genome differences across all comparisons
}
\examples{
differences <- as.data.frame(cbind(cellid=c(1:5), diff=runif(5, min=0, max=100)))
plot_diffdensity(differences)
}
