% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotClusters}
\alias{plotClusters}
\title{Plots a PCA reduction colored by cluster group}
\usage{
plotClusters(signs, palette, k)
}
\arguments{
\item{signs}{matrix with levels of exposure to each signature}

\item{palette}{vector with colors to use for each group. The length of this
vector must be equal to the number of cluster (k)}

\item{k}{number of clusters. It is recommended to estimate the
optimal number of clusters before}
}
\value{
A plot with samples clustered by similarity in signature exposition
}
\description{
This function draws a cluster plot with samples grouped by
exposure levels to ovarian copy-number signatures
}
