% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifydiff_functions.R
\name{unifySegments}
\alias{unifySegments}
\title{Alignment of genomic positions of segments}
\usage{
unifySegments(posSeg, data)
}
\arguments{
\item{posSeg}{dataframe with segments of the reference sample}

\item{data}{dataframe with segments of the experimental model}
}
\value{
dataframe with segments of the model aligned to those in reference
}
\description{
This function unifies boundaries of copy-number segments
}
