% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{CNconvert}
\alias{CNconvert}
\title{A helper function for CNPlot_events}
\usage{
CNconvert(e, base)
}
\arguments{
\item{e}{segment table with absolute copy numbers}

\item{base}{sizes of chromosomes}
}
\value{
segment table prepared for plotting
}
\description{
This function prepare profiles for plotting
}
