% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnsignature_functions.R
\name{extractCopynumberFeatures}
\alias{extractCopynumberFeatures}
\title{Extract features of copy number profiles}
\usage{
extractCopynumberFeatures(CN_data, cores = 1)
}
\arguments{
\item{CN_data}{segment table with copy numbers}

\item{cores}{cores to use. Default is 1}
}
\value{
A list with features per sample
}
\description{
This function extract feactures of CNAs.
Approach developed by Geoff Macintyre et al. 2018
}
\examples{
profiles<-getProfiles(segcn=cells_segcn,
    samples=unique(cells_segcn$sample)[1:4])
CN_features<-extractCopynumberFeatures(profiles)
}
