% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnsignature_functions.R
\name{getCentromereDistCounts}
\alias{getCentromereDistCounts}
\title{Get the breakpoint count per chromosome arm}
\usage{
getCentromereDistCounts(abs_profiles, centromeres, chrlen)
}
\arguments{
\item{abs_profiles}{absolute copy-number profiles}

\item{centromeres}{centromere positions}

\item{chrlen}{chromosome length}
}
\value{
Matrix with breakpoint count per chromosome arm
}
\description{
Helper function for extractCopynumberFeatures.
Approach developed by Geoff Macintyre et al. 2018
}
\examples{
profiles <- getProfiles(segcn=cells_segcn,
    samples=unique(cells_segcn$sample)[1:4])
gaps<-CNpare:::gaps
centromeres<-centromeres<-gaps[gaps[,8]=="centromere",]
bpchrarm<-getCentromereDistCounts(profiles,centromeres,CNpare:::chrlen)
}
