% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessing_functions.R
\name{getCINSamp}
\alias{getCINSamp}
\title{elect samples with CIN}
\usage{
getCINSamp(profiles)
}
\arguments{
\item{profiles}{list of profiles}
}
\value{
vector with sample names
}
\description{
This function filters out samples with <20 CNAs
}
\examples{
profiles <- getProfiles(segcn=cells_segcn,
    samples=unique(cells_segcn$sample))
cin.samples<-getCINSamp(profiles)
}
