% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessing_functions.R
\name{getCNbins.bin}
\alias{getCNbins.bin}
\title{Get copy-number in a bin}
\usage{
getCNbins.bin(b, pb, data, samples)
}
\arguments{
\item{b}{bin number}

\item{pb}{matrix with genomic positions of bins}

\item{data}{segment table of copy numbers of one bin}

\item{samples}{vector with sample names}
}
\value{
a list of copy-number values per bin
}
\description{
This is a helper function for transforming segment tables to bin tables
}
\examples{
posBins <- lapply(seq_len(22),function(chr)
    getBinsStartsEnds(window=500000, chr, lengthChr[chr]))
samples=unique(cells_segcn$sample)[1:2]
data=cells_segcn[cells_segcn$sample\%in\%samples,]
cn_bin <- getCNbins.bin(b=2,pb=data.table::rbindlist(posBins)[1:20],data,samples)
}
