% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessing_functions.R
\name{getCNbins.bin}
\alias{getCNbins.bin}
\title{Get copy-number in a bin}
\usage{
getCNbins.bin(b, pb, data)
}
\arguments{
\item{b}{bin number}

\item{pb}{matrix with genomic positions of bins}

\item{data}{segment table of copy numbers of one sample}
}
\value{
a list of copy-number values per bin
}
\description{
This is a helper function for transforming segment tables to bin tables
}
\examples{
posBins <- lapply(seq_len(22),function(chr)
    getBinsStartsEnds(window=500000, chr, lengthChr[chr]))
cell_bin <- getCNbins.bin(b=1,pb=data.table::rbindlist(posBins),
    data=cells_segcn[cells_segcn$sample=="22RV1",])
}
