% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessing_functions.R
\name{getCNbins.sample}
\alias{getCNbins.sample}
\title{Get copy-number per bins in a sample}
\usage{
getCNbins.sample(s, cn)
}
\arguments{
\item{s}{sample name}

\item{cn}{matrix with copy numbers in a bin (row) in samples (columns)}
}
\value{
a vector with copy numbers of samples in one bin
}
\description{
This is a helper function for transforming segment tables to bin tables
}
\examples{
posBins <- lapply(seq_len(22),function(chr)
    getBinsStartsEnds(window=500000, chr, lengthChr[chr]))
pb=data.table::rbindlist(posBins)[1:20]
samp=unique(cells_segcn$sample)[1:2]
data=data.table::as.data.table(cells_segcn[cells_segcn$sample\%in\%samp,c(5,4,1:3)])
data.table::setkey(pb, chromosome, start, end)
overlap=as.data.frame(data.table::foverlaps(data,pb))
cn=overlap[(overlap$chromosome=="1" & overlap$start==1 & overlap$end==500000),]
cell_bin=getCNbins.sample(s=samp[1],cn)
}
