% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocessing_functions.R
\name{getCNbins.sample}
\alias{getCNbins.sample}
\title{Get copy-number per bins in a sample}
\usage{
getCNbins.sample(posBins, data)
}
\arguments{
\item{posBins}{list with genomic positions of bins. Each list contains
data from each chromosome. Obtained from getBinsStartsEnds}

\item{data}{segment table of copy numbers of one sample}
}
\value{
bin table of copy numbers of one sample
}
\description{
This is a helper function for transforming segment tables to bin tables
}
\examples{
posBins <- lapply(seq_len(22),function(chr)
    getBinsStartsEnds(window=500000, chr, lengthChr[chr]))
cell_bin <- getCNbins.sample(posBins=posBins,
    data=cells_segcn[cells_segcn$sample=="22RV1",])
}
