% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifydiff_functions.R
\name{getDifference}
\alias{getDifference}
\title{Calculation of the extent of differences between two profiles}
\usage{
getDifference(unify)
}
\arguments{
\item{unify}{dataframe with copy numbers in segments of two profiles.
Obtained from unifySegments}
}
\value{
\% genome difference between two copy-number profiles
}
\description{
This function computes the \% genome difference
}
\examples{
unify<-unifySegments(posSeg=cells_segcn[cells_segcn$sample=="OVKATE",],
    data=cells_segcn[cells_segcn$sample=="OV-90",])
diff<-getDifference(unify)
}
