% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtool_functions.R
\name{getTopHit}
\alias{getTopHit}
\title{Get top hits}
\usage{
getTopHit(samples, measure, method = "all")
}
\arguments{
\item{samples}{vector with name of samples}

\item{measure}{dataframe with similarity metrics}

\item{method}{similarity metric to use for ordering . Default is all}
}
\value{
dataframe with similarity metrics for all comparisons
}
\description{
This function get the top hit for each comparison per sample
}
\examples{
posBins <- lapply(seq_len(22),function(chr)
    getBinsStartsEnds(window=500000, chr, lengthChr[chr]))
ccle_cn <- getCNbins(posBins=posBins, data=cells_segcn, samples=unique(cells_segcn$sample)[1:4])
exp_cell<-as.matrix(ccle_cn[,1])
colnames(exp_cell)<-unique(ccle_cn$sample)[1]
measures <- getSimilarities(dat1=exp_cell,dat2=ccle_cn)
tophits<-getTopHit(samples=unique(cells_segcn$sample)[1], measure=measures)
}
