% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtool_functions.R
\name{pair_euclidean}
\alias{pair_euclidean}
\title{Calculation of euclidean distance}
\usage{
pair_euclidean(cell, ccle, ccle.name)
}
\arguments{
\item{cell}{vector with bin-level copy number values of one sample}

\item{ccle}{vector with bin-level copy number values of model}

\item{ccle.name}{name of model}
}
\value{
Euclidean distance
}
\description{
This function calculates the Euclidean distance
}
