% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtool_functions.R
\name{pair_manhattan}
\alias{pair_manhattan}
\title{Calculation of manhattan distance}
\usage{
pair_manhattan(cell, ccle, ccle.name)
}
\arguments{
\item{cell}{vector with bin-level copy number values of one sample}

\item{ccle}{vector with bin-level copy number values of model}

\item{ccle.name}{name of model}
}
\value{
Manhattan distance
}
\description{
This function calculates the Manhattan distance
}
