% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtool_functions.R
\name{pair_manhattan}
\alias{pair_manhattan}
\title{Calculation of manhattan distance}
\usage{
pair_manhattan(cell, ccle, ccle.name, pvalue)
}
\arguments{
\item{cell}{vector with bin-level copy number values of one sample}

\item{ccle}{vector with bin-level copy number values of the cell model}

\item{ccle.name}{name of the cell model}

\item{pvalue}{a logical value to indicate if the empirical pvalue is calculated or not}
}
\value{
Manhattan distance
}
\description{
This function calculates the Manhattan distance
}
\examples{
posBins<-lapply(1,function(chr)
    getBinsStartsEnds(window=500000, chr, lengthChr[chr]))
ccle_cn<-getCNbins(posBins=posBins, data=cells_segcn, samples=unique(cells_segcn$sample)[1:2])
exp_cell<-as.matrix(ccle_cn[,1])
mod_cell<-as.matrix(ccle_cn[,2])
m<-pair_manhattan(exp_cell,mod_cell,unique(cells_segcn$sample)[2],pvalue=FALSE)
}
