% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtool_functions.R
\name{pair_pearson}
\alias{pair_pearson}
\title{Perform a pearson correlation test}
\usage{
pair_pearson(cell, ccle, ccle.name, pvalue)
}
\arguments{
\item{cell}{vector with bin-level copy number values of one sample}

\item{ccle}{vector with bin-level copy number values of the cell model}

\item{ccle.name}{name of the cell model}

\item{pvalue}{a logical value to indicate if the empirical pvalue is calculated or not}
}
\value{
Pearson's r
}
\description{
This function calculates the Pearson's r
}
\examples{
posBins<-lapply(1,function(chr)
    getBinsStartsEnds(window=1000000, chr, lengthChr[chr]))
ccle_cn<-getCNbins(posBins=posBins, data=cells_segcn, samples=unique(cells_segcn$sample)[1:2])
exp_cell<-as.matrix(ccle_cn[,1])
mod_cell<-as.matrix(ccle_cn[,2])
r<-pair_pearson(exp_cell,mod_cell,unique(cells_segcn$sample)[2],pvalue=FALSE)
}
