% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotClusters}
\alias{plotClusters}
\title{Plots a PCA reduction colored by cluster group}
\usage{
plotClusters(matrix, palette, k)
}
\arguments{
\item{matrix}{matrix with samples in rows and variables in columns}

\item{palette}{vector with colors to use for each group. The length of this
vector must be equal to the number of cluster (k)}

\item{k}{number of clusters. It is recommended to estimate the
optimal number of clusters before}
}
\value{
A plot with samples clustered by similarity in signature exposition
}
\description{
This function draws a scatter plot with samples grouped and
colored according to its cluster assignment
}
\examples{
matrix <- cbind(s1=runif(10, min=0, max=1), s2=runif(10, min=0, max=1),
    s3=runif(10, min=0, max=1))
plotClusters(matrix, palette=c("#2E9FDF", "#E7B800"), k=2)
}
