% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotClusters}
\alias{plotClusters}
\title{Plots a scatter plots with samples clustered}
\usage{
plotClusters(matrix, samples)
}
\arguments{
\item{matrix}{matrix variables in rows and samples in columns}

\item{samples}{vector with samples in the closest cluster of the input sample}
}
\value{
A plot with samples clustered by similarity in signature exposition
}
\description{
This function draws a scatter plot with samples grouped based of signature
exposures. Only the closest cluster to the input sample is colored.
}
\examples{
matrix<-cbind(s1=runif(10, min=0, max=1), s2=runif(10, min=0, max=1),
    s3=runif(10, min=0, max=1))
rownames(matrix)<-paste0("c", seq(1,10))
cell<-matrix(runif(n = 3, min = 0, max = 1))
samples<-getClusterSamples(t(matrix), cell, include=FALSE)
plotClusters(t(matrix), samples)
}
