% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnsignature_functions.R
\name{quantifySignatures}
\alias{quantifySignatures}
\title{Quantifying signatures}
\usage{
quantifySignatures(sample_by_component, component_by_signature = NULL)
}
\arguments{
\item{sample_by_component}{matrix with components per sample}

\item{component_by_signature}{matrix with components per signature}
}
\value{
A matrix with signature exposures per sample
}
\description{
This function quantify copy-number signatures.
Approach developed by Geoff Macintyre et al. 2018
}
\examples{
\dontrun{
profiles <- getProfiles(segcn=cells_segcn,
    samples=unique(cells_segcn$sample)[1:4])
CN_features<-extractCopynumberFeatures(profiles)
sample_by_component <- generateSampleByComponentMatrix(CN_features,
    all_components=component_parameters)
signature_quantification <- quantifySignatures(sample_by_component)
}
}
