% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifydiff_functions.R
\name{unifySegments}
\alias{unifySegments}
\title{Alignment of genomic positions of segments}
\usage{
unifySegments(posSeg, data, window = 1e+05)
}
\arguments{
\item{posSeg}{dataframe with segments of the reference sample}

\item{data}{dataframe with segments of the experimental model}

\item{window}{boundary margin for smoothing segments. Default is 100000}
}
\value{
dataframe with segments of the model aligned to those in reference
}
\description{
This function unifies boundaries of copy-number segments.
Before to apply this function, please double-check that both profiles include
the same chromosomes
}
\examples{
unify<-unifySegments(posSeg=cells_segcn[cells_segcn$sample=="OVKATE",],
    data=cells_segcn[cells_segcn$sample=="OV-90",])
}
